eu-78 delete useless folder n files, add slashs to urls, update imports
