--- apps/core/schemas/role.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-class AvailableRolesOut(BaseModel):
-    role_id: int
-    role_name: str
