--- apps/hostel/temp/hostel_list_view.py
@@ -1,16 +1,0 @@
-from sqlalchemy import (MetaData, Column, Table, Integer, VARCHAR)
-
-
-metadata_obj = MetaData()
-
-
-hostel_list_view = Table('hostel_list_view', metadata_obj,
-          Column('university_id', Integer),
-          Column('hostel_id', Integer),
-          Column('number', Integer),
-          Column('name', VARCHAR(100)),
-          Column('city', VARCHAR(100)),
-          Column('street', VARCHAR(100)),
-          Column('build', VARCHAR(10)),
-          Column('commandant_id', Integer),
-          Column('commandant_full_name', VARCHAR(255)))
