+++ apps/templates/hostel_booking_template.docx
