--- tags_metadata.py
@@ -1,34 +1,0 @@
-metadata = [
-    {
-        "name": "Authorization",
-        "description": "Operations with signin, signup of user. Also included checking student existanse and changing password endpoint.",
-        "externalDocs": {
-            "description": "See epic for more info",
-            "url": "https://euniversity.atlassian.net/browse/EU-12",
-        },
-    },
-    {
-        "name": "Student dashboard",
-        "description": "All endpoints avaliable for user with role 'Student'.",
-        "externalDocs": {
-            "description": "See epic for more info",
-            "url": "https://euniversity.atlassian.net/browse/EU-15",
-        },
-    },
-    {
-        "name": "Admin dashboard",
-        "description": "All endpoints avaliable for user with role 'Admin'.",
-        "externalDocs": {
-            "description": "See epic for more info",
-            "url": "https://euniversity.atlassian.net/browse/EU-16",
-        },
-    },
-    {
-        "name": "SuperAdmin dashboard",
-        "description": "All endpoints avaliable for user with role 'Super Admin'.",
-        "externalDocs": {
-            "description": "See epic for more info",
-            "url": "https://euniversity.atlassian.net/browse/EU-17",
-        },
-    }
-]
